@@ -5,12 +5,8 @@ export function get() {
 }
 
 export async function update(settings) {
-  const res = await fetchURL(`/api/settings`, {
+  await fetchURL(`/api/settings`, {
     method: "PUT",
     body: JSON.stringify(settings),
   });
-
-  if (res.status !== 200) {
-    throw new Error(res.status);
-  }
 }
