@@ -1,8 +1,9 @@
 import store from "@/store";
-import { renew } from "@/utils/auth";
+import { renew, logout } from "@/utils/auth";
 import { baseURL } from "@/utils/constants";
+import { encodePath } from "@/utils/url";
 
-export async function fetchURL(url, opts) {
+export async function fetchURL(url, opts, auth = true) {
   opts = opts || {};
   opts.headers = opts.headers || {};
 
@@ -17,12 +18,26 @@ export async function fetchURL(url, opts) {
       },
       ...rest,
     });
-  } catch (error) {
-    return { status: 0 };
+  } catch {
+    const error = new Error("000 No connection");
+    error.status = 0;
+
+    throw error;
   }
 
-  if (res.headers.get("X-Renew-Token") === "true") {
+  if (auth && res.headers.get("X-Renew-Token") === "true") {
     await renew(store.state.jwt);
+  }
+
+  if (res.status < 200 || res.status > 299) {
+    const error = new Error(await res.text());
+    error.status = res.status;
+
+    if (auth && res.status == 401) {
+      logout();
+    }
+
+    throw error;
   }
 
   return res;
@@ -44,4 +59,19 @@ export function removePrefix(url) {
   if (url === "") url = "/";
   if (url[0] !== "/") url = "/" + url;
   return url;
+}
+
+export function createURL(endpoint, params = {}, auth = true) {
+  const url = new URL(encodePath(endpoint), origin + baseURL);
+
+  const searchParams = {
+    ...(auth && { auth: store.state.jwt }),
+    ...params,
+  };
+
+  for (const key in searchParams) {
+    url.searchParams.set(key, searchParams[key]);
+  }
+
+  return url.toString();
 }
