@@ -1,6 +1,6 @@
 <template>
   <div>
-    <div id="progress">
+    <div v-if="progress" class="progress">
       <div v-bind:style="{ width: this.progress + '%' }"></div>
     </div>
     <sidebar></sidebar>
@@ -9,6 +9,7 @@
       <shell v-if="isExecEnabled && isLogged && user.perm.execute" />
     </main>
     <prompts></prompts>
+    <upload-files></upload-files>
   </div>
 </template>
 
@@ -17,6 +18,7 @@ import { mapState, mapGetters } from "vuex";
 import Sidebar from "@/components/Sidebar";
 import Prompts from "@/components/prompts/Prompts";
 import Shell from "@/components/Shell";
+import UploadFiles from "../components/prompts/UploadFiles";
 import { enableExec } from "@/utils/constants";
 
 export default {
@@ -25,6 +27,7 @@ export default {
     Sidebar,
     Prompts,
     Shell,
+    UploadFiles,
   },
   computed: {
     ...mapGetters(["isLogged", "progress"]),
