@@ -4,15 +4,13 @@
       <action icon="close" :label="$t('buttons.close')" @action="close()" />
       <title>{{ req.name }}</title>
 
-      <template #actions>
-        <action
-          v-if="user.perm.modify"
-          id="save-button"
-          icon="save"
-          :label="$t('buttons.save')"
-          @action="save()"
-        />
-      </template>
+      <action
+        v-if="user.perm.modify"
+        id="save-button"
+        icon="save"
+        :label="$t('buttons.save')"
+        @action="save()"
+      />
     </header-bar>
 
     <breadcrumbs base="/files" noLink />
