@@ -7,9 +7,9 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/asdine/storm"
-	toml "github.com/pelletier/go-toml"
-	yaml "gopkg.in/yaml.v2"
+	"github.com/asdine/storm/v3"
+	"github.com/pelletier/go-toml/v2"
+	"gopkg.in/yaml.v2"
 
 	"github.com/filebrowser/filebrowser/v2/auth"
 	"github.com/filebrowser/filebrowser/v2/settings"
